Updated a number of files to their newest versions and added new files.

+++ NewsComponents/Storage/FileCacheManager.cs
@@ -1,16 +1,39 @@
 #region CVS Version Header
 /*
- * $Id: FileCacheManager.cs,v 1.14 2005/04/04 03:53:01 haacked Exp $
- * Last modified by $Author: haacked $
- * Last modified at $Date: 2005/04/04 03:53:01 $
- * $Revision: 1.14 $
+ * $Id: FileCacheManager.cs,v 1.29 2007/02/17 14:45:53 t_rendelmann Exp $
+ * Last modified by $Author: t_rendelmann $
+ * Last modified at $Date: 2007/02/17 14:45:53 $
+ * $Revision: 1.29 $
+ */
+#endregion
+
+#region CVS Version Log
+/*
+ * $Log: FileCacheManager.cs,v $
+ * Revision 1.29  2007/02/17 14:45:53  t_rendelmann
+ * switched: Resource.Manager indexer usage to strongly typed resources (StringResourceTool)
+ *
+ * Revision 1.28  2007/02/10 17:22:50  carnage4life
+ * Added code to handle FileNotFoundException in LuceneIndexModifier. We now reset the index when this occurs because it indicates that the search index has been corrupted.
+ *
+ * Revision 1.27  2006/09/29 00:50:48  carnage4life
+ * Switched to using FileHelper for loading .bin files
+ *
+ * Revision 1.26  2006/09/03 19:08:50  carnage4life
+ * Added support for favicons
+ *
+ * Revision 1.25  2006/08/18 19:10:57  t_rendelmann
+ * added an "id" XML attribute to the feedsFeed. We need it to make the feed items (feeditem.id + feed.id) unique to enable progressive indexing (lucene)
+ *
  */
 #endregion
 
 using System;
+using System.Collections;
 using System.IO;
 using System.Xml;
 using NewsComponents.Feed;
+using NewsComponents.Resources;
 using NewsComponents.Utils;
 
 namespace NewsComponents.Storage {	
@@ -38,45 +61,21 @@ namespace NewsComponents.Storage {
 			if(Directory.Exists(cacheDirectory)){
 				this.cacheDirectory = cacheDirectory; 
 			}else{
-				throw new IOException(Resource.Manager["RES_ExceptionDirectoryNotExistsMessage", cacheDirectory]); 
+				throw new IOException(ComponentsText.ExceptionDirectoryNotExistsMessage(cacheDirectory)); 
 			}
-		}		
+		}				
 		
-		
-		/* 
-		/// <summary>
-		///  Sets the maximum amount of time an item should be kept in the 
-		/// cache for a particular feed. This overrides the value of the 
-		/// maxItemAge property. 
-		/// </summary>
-		/// <remarks>This method should be thread-safe</remarks>
-		/// <param name="feed">The feed</param>
-		/// <param name="age">The maximum amount of time items should be kept for the 
-		/// specified feed.</param>
-		/// <exception cref="IOException">If an error occurs while saving the feed</exception>
-		[MethodImpl(MethodImplOptions.Synchronized)]
-		[Obsolete("Should live in feedlist.xml")]
-		public override void SetMaxItemAge(feedsFeed feed, TimeSpan age){
-			FeedDetailsInternal fd = this.GetFeed(feed);
-			if(fd != null){ //perhaps should throw exception if null? Not much for caller to do though. 
-				fd.MaxItemAge = age; 
-				feed.cacheurl = this.SaveFeed(fd); 
-			}
-		} 
 
 		/// <summary>
-		/// Gets the maximum amount of time an item is kept in the 
-		/// cache for a particular feed. 
+		/// Returns the directory path to the Cache.
 		/// </summary>
-		/// <param name="feed">The feed identifier</param>
-		/// <exception cref="IOException">If an error occurs while retrieving the feed</exception>
-		[Obsolete("Should live in feedlist.xml")]
-		public override TimeSpan GetMaxItemAge(feedsFeed feed){
-			 
-			 //TODO: Implement more efficiently by using XmlReader to scan XML on disk
-			 FeedDetails fd   = this.GetFeed(feed);
-			 return (fd == null ? this.maxItemAge : fd.MaxItemAge); 
-		} */
+		internal override string CacheLocation {
+			get{ 
+				return this.cacheDirectory; 
+			}		
+		}
+
+
 
 		/// <summary>
 		/// Returns an RSS feed. 
@@ -88,7 +87,7 @@ namespace NewsComponents.Storage {
 			if (null == feed || null == feed.cacheurl)
 				return null;
 
-			FeedDetailsInternal fi = null; 
+			FeedInfo fi = null; 
 
 			string cachelocation = Path.Combine(this.cacheDirectory, feed.cacheurl);
 
@@ -97,7 +96,7 @@ namespace NewsComponents.Storage {
 				using (Stream feedStream = FileHelper.OpenForRead(cachelocation)) {
 					fi = RssParser.GetItemsForFeed(feed, feedStream, true); 
 				}				  			  				
-
+				this.LoadItemContent(fi); 
 			}
 
 			return fi; 			
@@ -112,36 +111,60 @@ namespace NewsComponents.Storage {
 		/// <returns>An identifier for the saved feed. </returns>		
 		internal override string SaveFeed(FeedDetailsInternal feed){
 			
-			string feedLocation = null;
-
-			if (feed.Type == FeedType.Rss) {
+			string feedLocation = feed.FeedLocation;
+			
+//			if (feed.Type == FeedType.Rss) {
 
 				lock(this){
 
 					if((feed.FeedLocation == null) || (feed.FeedLocation.Length == 0)){
-						feed.FeedLocation = GetCacheUrlName(new Uri(feed.Link)); 
-					}
-
-					feedLocation= feed.FeedLocation;
+						feed.FeedLocation = feedLocation = GetCacheUrlName(feed.Id, new Uri(feed.Link)); 
+					}					
 				}				
-			
-				// RemoveExpiredItems(feed.ItemsList, maxItemAge);
-			
+		
+				//get location of binary file containing RSS item contents 
+				string feedContentLocation = feedLocation.Substring(0, feedLocation.Length - 4) + ".bin";								
+
+				//write main RSS feed
 				using (MemoryStream stream = new MemoryStream()) {
 					XmlTextWriter writer = new XmlTextWriter(new StreamWriter(stream )); 
-					feed.WriteTo(writer); 
+					feed.WriteTo(writer,true); 
 					writer.Flush(); 
 					FileHelper.WriteStreamWithRename(Path.Combine(this.cacheDirectory,feedLocation), stream);
 				}
 			
-			} else if (feed.Type == FeedType.Nntp) {
+				//write binary file containing RSS item contents 
+				using (MemoryStream stream = new MemoryStream()) {
+					FileStream fs = null; 
+					BinaryReader reader = null; 
+					BinaryWriter writer = new BinaryWriter(stream);
+					
+					try{ 
+						if(File.Exists(Path.Combine(this.cacheDirectory, feedContentLocation))){
+							fs = new FileStream(Path.Combine(this.cacheDirectory, feedContentLocation), FileMode.OpenOrCreate);
+							reader = new BinaryReader(fs);
+						}					 
+						feed.WriteItemContents(reader, writer); 												
+						writer.Write(FileHelper.EndOfBinaryFileMarker); 
+						writer.Flush(); 						
+					}finally{
+						if(reader != null){
+							reader.Close(); 
+							fs.Close(); 
+						}
+					}
 
-				throw new NotImplementedException("NntpInfo class impl. missing");
-				
+					FileHelper.WriteStreamWithRename(Path.Combine(this.cacheDirectory,feedContentLocation), stream);									
+				}//using(MemoryStream...) {
 
-			} else {
-				throw new InvalidOperationException("Unknown/unhandled FeedDetails impl.");
-			}
+//			} else if (feed.Type == FeedType.Nntp) {
+//
+//				throw new NotImplementedException("NntpInfo class impl. missing");
+//				
+//
+//			} else {
+//				throw new InvalidOperationException("Unknown/unhandled FeedDetails impl.");
+//			}								
 
 			return feedLocation;
 		} 
@@ -153,13 +176,22 @@ namespace NewsComponents.Storage {
 		/// <param name="feed">The feed to remove</param>
 		public override void RemoveFeed(feedsFeed feed){
 
+			if (feed == null || feed.cacheurl == null)
+				return;
+
 			string cachelocation = Path.Combine(this.cacheDirectory, feed.cacheurl);
-		
+			string feedContentLocation = Path.Combine(this.cacheDirectory, 
+											feed.cacheurl.Substring(0, feed.cacheurl.Length - 4) + ".bin");											
+				
 			try{ 
 
-			if(File.Exists(cachelocation)){
-				FileHelper.Delete(cachelocation); 
-			}
+				if(File.Exists(cachelocation)){
+					FileHelper.Delete(cachelocation); 
+				}
+
+				if(File.Exists(feedContentLocation)){
+					FileHelper.Delete(feedContentLocation); 
+				}		
 
 			}catch(IOException iox){ 
 				_log.Debug("RemoveFeed: Could not delete " + cachelocation, iox); 
@@ -200,17 +232,127 @@ namespace NewsComponents.Storage {
 			return File.Exists(cachelocation); 
 		}
 
+
+		/// <summary>
+		/// Loads the contents of the NewsItem from the cache. The provided NewsItem must have 
+		/// non-null value for its Id property. 
+		/// </summary>
+		/// <param name="item"></param>
+		public override void LoadItemContent(NewsItem item){		
+
+			FileStream fs = null; 
+			BinaryReader reader = null; 
+
+			try{ 
+
+				string feedContentLocation = item.Feed.cacheurl.Substring(0, item.Feed.cacheurl.Length - 4) + ".bin";											
+
+				if(File.Exists(Path.Combine(this.cacheDirectory, feedContentLocation))){
+
+					fs = FileHelper.OpenForRead(Path.Combine(this.cacheDirectory, feedContentLocation));					
+					reader = new BinaryReader(fs); 
+
+					string id = reader.ReadString(); 
+				
+					while(!id.Equals(FileHelper.EndOfBinaryFileMarker)){
+						int count = reader.ReadInt32();
+						byte[] content = reader.ReadBytes(count);
+				
+						if(item.Id.Equals(id)){							
+							item.SetContent(content, ContentType.Html); 
+							item.RefreshRelationCosmos();
+							break; 
+						}
+						id = reader.ReadString(); 
+					}//while(!id.Equals(...))
+				}
+			
+			}finally{
+				if(reader != null) {
+					reader.Close(); 
+				}
+
+				if(fs != null){
+					fs.Close(); 
+				}			
+			}//finally 
+
+		}
+
+
+		/// <summary>
+		/// Loads the content of the unread NewsItems from the binary file 
+		/// where item contents are contained. This is a memory-saving performance 
+		/// optimization so we only have the content of items that are unread on load.  
+		/// </summary>
+		/// <param name="fi"></param>
+		private void LoadItemContent(FeedInfo fi){
+					
+			Hashtable unreadItems = new Hashtable(); 
+			FileStream fs = null; 
+			BinaryReader reader = null; 
+
+
+			//get list of unread items 
+			foreach(NewsItem item in fi.itemsList){
+				if(!item.BeenRead){
+					try{
+						unreadItems.Add(item.Id, item); 
+					}catch(ArgumentException){
+						/* we don't test using ContainsKey() before Add() for performance reasons */
+					}
+				}
+			}
+
+			try{ 
+
+				string feedContentLocation = fi.feedLocation.Substring(0, fi.feedLocation.Length - 4) + ".bin";											
+
+				if(File.Exists(Path.Combine(this.cacheDirectory, feedContentLocation))){
+
+					fs = FileHelper.OpenForRead(Path.Combine(this.cacheDirectory, feedContentLocation));
+					reader = new BinaryReader(fs); 
+
+					string id = reader.ReadString(); 
+				
+					while(!id.Equals(FileHelper.EndOfBinaryFileMarker)){
+						int count = reader.ReadInt32();
+						byte[] content = reader.ReadBytes(count);
+				
+						if(unreadItems.Contains(id)){
+							NewsItem ni = (NewsItem) unreadItems[id];  
+							ni.SetContent(content, ContentType.Html); 
+							ni.RefreshRelationCosmos();
+						}
+						id = reader.ReadString(); 
+					}//while(!id.Equals(...))
+				}
+			
+			}finally{
+				if(reader != null) {
+					reader.Close(); 
+				}
+
+				if(fs != null){
+					fs.Close(); 
+				}			
+			}//finally 
+			 		
+		}	
+		
+
 		/// <summary>
 		/// Get a file name that the file can be cached.
 		/// </summary>
+		/// <param name="id">Id to be used to build</param>
 		/// <param name="uri">The uri of the rss document.</param>
 		/// <returns>a filename that may be used to save the cached file.</returns>
-		private static string GetCacheUrlName(Uri uri) {
+		private static string GetCacheUrlName(string id, Uri uri) {
 			string path = null;
 			if (uri.IsFile || uri.IsUnc) {
-				path = uri.GetHashCode() + "." + System.Guid.NewGuid() + ".xml";
+				path = uri.GetHashCode() + "." + id + ".xml";
 			}else{
-				path = uri.Host + "." + uri.Port + "." + uri.GetHashCode() + "." + System.Guid.NewGuid() + ".xml";
+				path = uri.Host + "." + uri.Port + "." + uri.GetHashCode() + "." + id + ".xml";
 			}
 			return path.Replace("-","");
 		}
